refactor(library/Byte): renames "index.ts" to "exports.ts"
